requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
drf-spectacular==0.22.0
importlib-resources==5.7.1
inflection==0.5.1
jsonschema==4.4.0
psycopg2==2.9.3
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
zipp==3.8.0
